requirements.txt
absl-py==0.9.0
aiohttp==3.5.4
altgraph==0.16.1
aniso8601==8.0.0
apiclient==1.0.4
astor==0.8.1
astroid==2.4.2
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bottle==0.12.17
bottle-websocket==0.2.9
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dnspython==1.16.0
Eel==0.10.4
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
future==0.17.1
gast==0.2.2
gevent==1.4.0
gevent-websocket==0.10.1
google-api-core==1.21.0
google-api-python-client==1.7.9
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.13.0
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.3.4
macholib==1.11
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
modulegraph==0.17
mongoengine==0.20.0
mpmath==1.1.0
multidict==4.5.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
oauth2client==4.1.3
opt-einsum==3.2.0
pandocfilters==1.4.2
parso==0.5.1
passlib==1.7.2
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyGetWindow==0.0.6
Pygments==2.4.2
PyInstaller==3.5
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.10.1
PyMsgBox==1.0.7
pyparsing==2.4.0
PyRect==0.1.4
pyrsistent==0.15.3
PyScreeze==0.1.21
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.13.0
python-http-client==3.2.7
PyTweening==1.0.3
pytz==2019.1
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.0.2
requests==2.22.0
rsa==4.0
scipy==1.3.0
Send2Trash==1.5.0
sendgrid==6.4.1
sh==1.12.14
six==1.12.0
soupsieve==1.9.1
starkbank-ecdsa==1.0.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.1
tornado==6.0.3
traitlets==4.3.2
twilio==6.36.0
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.7.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
whichcraft==0.6.0
widgetsnbextension==3.5.0
wrapt==1.12.1
WTForms==2.3.1
yarl==1.3.0
